Render path for crews
